move files into components folder
