+++ oyster/src/components/App.js
@@ -1,6 +1,6 @@
-import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
+import React, { Component } from "react";
+import logo from "assets/images/logo.svg";
+import "./App.css";
 
 class App extends Component {
   render() {
